Observer Pattern을 위한 인터페이스 추가 외
Observer Pattern을 위한 인터페이스 추가
Health와 관련된 처리를 Observer Pattern으로 처리

+++ Prepper/Source/Prepper/Interfaces/PlayerAbility.h
@@ -27,7 +27,7 @@ public:
 	virtual void UseQuickSlotItem(int Idx) PURE_VIRTUAL();
 	virtual void EquipWeapon(class AWeaponActor* Weapon) PURE_VIRTUAL();
 
-	virtual void Heal(int Amount) PURE_VIRTUAL();
-	virtual void Eat(int Amount) PURE_VIRTUAL();
-	virtual void Drink(int Amount) PURE_VIRTUAL();
+	virtual void Heal(float Amount) PURE_VIRTUAL();
+	virtual void Eat(float Amount) PURE_VIRTUAL();
+	virtual void Drink(float Amount) PURE_VIRTUAL();
 };
